requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
bcrypt==4.1.3
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.113.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nodeenv==1.8.0
orjson==3.10.3
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
setuptools==69.2.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tomlkit==0.12.4
typer==0.12.3
typing_extensions==4.10.0
ujson==5.10.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
